Divide admin panel code to usercontrols
