+++ SchoolManagerWPF/View/TabPages/Classes.xaml
@@ -1,0 +1,130 @@
+﻿<UserControl x:Class="SchoolManagerWPF.View.TabPages.Classes"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:local="clr-namespace:SchoolManagerWPF.View.TabPages"
+             xmlns:resx="clr-namespace:SchoolManagerModel.Resources;assembly=SchoolManagerModel"
+             xmlns:controls="clr-namespace:SchoolManagerWPF.View.Controls"
+             xmlns:converters="clr-namespace:SchoolManagerWPF.Converters"
+             mc:Ignorable="d"
+             d:DesignHeight="450"
+             d:DesignWidth="800">
+    <UserControl.Resources>
+        <BooleanToVisibilityConverter x:Key="BooleanToVisibilityConverter" />
+        <converters:RoleToIntConverter x:Key="RoleToIntConverter" />
+        <converters:StringToVisibilityConverter x:Key="StringToVisibilityConverter" />
+    </UserControl.Resources>
+    <Grid Background="#FFE5E5E5"
+          Name="ClassesGrid"
+          Margin="30,30,30,30">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="auto"></RowDefinition>
+            <RowDefinition Height="30px"></RowDefinition>
+            <RowDefinition Height="auto"></RowDefinition>
+            <RowDefinition Height="auto"></RowDefinition>
+            <RowDefinition Height="15px"></RowDefinition>
+            <RowDefinition Height="auto"></RowDefinition>
+            <RowDefinition Height="auto"></RowDefinition>
+            <RowDefinition Height="auto"></RowDefinition>
+            <RowDefinition Height="auto"></RowDefinition>
+        </Grid.RowDefinitions>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="auto"></ColumnDefinition>
+            <ColumnDefinition Width="30"></ColumnDefinition>
+            <ColumnDefinition Width="auto"></ColumnDefinition>
+            <ColumnDefinition Width="auto"></ColumnDefinition>
+            <ColumnDefinition Width="auto"></ColumnDefinition>
+            <ColumnDefinition Width="30px"></ColumnDefinition>
+            <ColumnDefinition Width="auto"></ColumnDefinition>
+            <ColumnDefinition Width="auto"></ColumnDefinition>
+
+        </Grid.ColumnDefinitions>
+        <DataGrid ItemsSource="{Binding Classes, ValidatesOnDataErrors=True}"
+                  AutoGenerateColumns="False"
+                  Height="auto"
+                  CanUserDeleteRows="False"
+                  CanUserAddRows="False"
+                  IsReadOnly="True"
+                  Grid.Row="0"
+                  Grid.ColumnSpan="7">
+            <DataGrid.Columns>
+                <DataGridTextColumn Header="{x:Static resx:UI.ID}"
+                                    Binding="{Binding Id}"
+                                    IsReadOnly="True" />
+                <DataGridTextColumn Header="{x:Static resx:UI.Name}"
+                                    Binding="{Binding Name, ValidatesOnDataErrors=True}" />
+            </DataGrid.Columns>
+        </DataGrid>
+        <Label Content="{x:Static resx:UI.ClassYear}"
+               ContentStringFormat="{}{0}:"
+               Grid.Row="2"
+               Grid.Column="0" />
+        <TextBox Text="{Binding ClassYear, Mode=TwoWay}"
+                 Grid.Row="2"
+                 Grid.Column="2"
+                 MinWidth="100px"
+                 VerticalAlignment="Center" />
+        <Grid Grid.Row="3"
+              Grid.Column="0"
+              Grid.ColumnSpan="4"
+              Width="auto"
+              Height="auto"
+              Name="ClassYearValidationErrorsGrid"
+              Visibility="{Binding ClassYearValidationErrors, Converter={StaticResource StringToVisibilityConverter}}">
+
+            <Label VerticalAlignment="Top"
+                   FontWeight="Bold"
+                   HorizontalAlignment="Stretch"
+                   Width="{Binding ActualWidth, ElementName=ClassYearValidationErrorsGrid, Mode=OneWay}">
+
+                <TextBlock TextWrapping="Wrap"
+                           Text="{Binding ClassYearValidationErrors}"
+                           HorizontalAlignment="Stretch"
+                           Background="{x:Null}"
+                           Foreground="Red" />
+            </Label>
+        </Grid>
+        <Button Content="{x:Static resx:UI.AddClass}"
+                HorizontalAlignment="Center"
+                VerticalAlignment="Center"
+                Command="{Binding AddClassCommand}"
+                Grid.Row="2"
+                Grid.Column="6"
+                Grid.RowSpan="4"
+                Height="65"
+                MinWidth="150px"
+                Width="auto" />
+        <Label Content="{x:Static resx:UI.Class}"
+               ContentStringFormat="{}{0}:"
+               Grid.Row="5"
+               Grid.Column="0" />
+        <TextBox Text="{Binding Class}"
+                 Grid.Row="5"
+                 Grid.Column="2"
+                 MinWidth="100px"
+                 VerticalAlignment="Center" />
+        <Grid Grid.Row="7"
+              Grid.Column="0"
+              Grid.ColumnSpan="4"
+              Width="auto"
+              Height="auto"
+              Name="ClassValidationErrorsGrid"
+              Visibility="{Binding ClassValidationErrors, Converter={StaticResource StringToVisibilityConverter}}">
+
+            <Label VerticalAlignment="Top"
+                   FontWeight="Bold"
+                   HorizontalAlignment="Stretch"
+                   Width="{Binding ActualWidth, ElementName=ClassValidationErrorsGrid, Mode=OneWay}"
+                   Height="auto">
+
+                <TextBlock TextWrapping="Wrap"
+                           Text="{Binding ClassValidationErrors}"
+                           HorizontalAlignment="Stretch"
+                           Background="{x:Null}"
+                           Foreground="Red" />
+            </Label>
+        </Grid>
+
+    </Grid>
+</UserControl>
